Renamed Http to Fitch
Renamed class to suit package name. Added missing packages and added lint script
