+++ .gitignore
@@ -12,4 +12,4 @@ node_modules
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-lib
+dist
